+++ src/services/ContoService.js
@@ -1,16 +1,16 @@
-import { BASE_URL, QUIZZES_EP, QUIZZE_EP } from './ApiURI';
+import { BASE_URL, CONTOS_EP, CONTO_EP } from './ApiURI';
 import Service from './Service';
 
 
 
-class QuizzService extends Service {
+class ContoService extends Service {
 
-    async save(quizz) {
+    async save(conto) {
 
-        return  fetch(`${BASE_URL}${QUIZZES_EP}`, {
+        return  fetch(`${BASE_URL}${CONTOS_EP}`, {
             method: "POST",
             headers: { "Content-Type": "application/json"},
-            body: JSON.stringify(quizz)
+            body: JSON.stringify(conto)
         })
         .then(res => {
             this.isFailure(res);
@@ -24,7 +24,7 @@ class QuizzService extends Service {
     }
 
     get list() {
-        return fetch(`${BASE_URL}${QUIZZES_EP}`)
+        return fetch(`${BASE_URL}${CONTOS_EP}`)
             .then(res => {
                 this.isFailure(res);
                 this.isNotFound(res, "Ops... não foi possível encontrar o caminho solicitado");
@@ -37,7 +37,7 @@ class QuizzService extends Service {
     }
 
     async findById(id) {
-        return fetch(`${BASE_URL}${QUIZZES_EP}/${QUIZZE_EP}?id=${id}`)
+        return fetch(`${BASE_URL}${CONTOS_EP}/${CONTO_EP}?id=${id}`)
             .then(res => {
                 this.isFailure(res);
                 this.isNotFound(res, `Ops... não foi possível encontrar o caminho solicitado / status code: ${res.status}`);
@@ -51,4 +51,4 @@ class QuizzService extends Service {
 
 }
 
-export default QuizzService;
+export default ContoService;
