+++ src/services/QuizzService.js
@@ -1,16 +1,16 @@
-import { BASE_URL, CONTOS_EP, CONTO_EP } from './ApiURI';
+import { BASE_URL, QUIZZ_EP,QUIZZES_EP} from './ApiURI';
 import Service from './Service';
 
 
 
-class ContoService extends Service {
+class QuizzService extends Service {
 
-    async save(conto) {
+    async save(quizz) {
 
-        return  fetch(`${BASE_URL}${CONTOS_EP}`, {
+        return  fetch(`${BASE_URL}${QUIZZES_EP}`, {
             method: "POST",
             headers: { "Content-Type": "application/json"},
-            body: JSON.stringify(conto)
+            body: JSON.stringify(quizz)
         })
         .then(res => {
             this.isFailure(res);
@@ -19,12 +19,12 @@ class ContoService extends Service {
             return res.json();
         })
         .catch(error => {
-            throw new Error("Houve um erro durante a tentativa de salvar o conto " + error.message);
+            throw new Error("Houve um erro durante a tentativa de salvar o quizz " + error.message);
         });
     }
 
     get list() {
-        return fetch(`${BASE_URL}${CONTOS_EP}`)
+        return fetch(`${BASE_URL}${QUIZZES_EP}`)
             .then(res => {
                 this.isFailure(res);
                 this.isNotFound(res, "Ops... não foi possível encontrar o caminho solicitado");
@@ -32,12 +32,12 @@ class ContoService extends Service {
                 return res.json();
             })
             .catch(error => {
-                throw new Error("Houve um erro durante ao buscar a lista de contos " + error.message);
+                throw new Error("Houve um erro durante ao buscar a lista de quizzes " + error.message);
             });
     }
 
     async findById(id) {
-        return fetch(`${BASE_URL}${CONTOS_EP}/${CONTO_EP}?id=${id}`)
+        return fetch(`${BASE_URL}${QUIZZES_EP}/${QUIZZ_EP}?id=${id}`)
             .then(res => {
                 this.isFailure(res);
                 this.isNotFound(res, `Ops... não foi possível encontrar o caminho solicitado / status code: ${res.status}`);
@@ -51,4 +51,4 @@ class ContoService extends Service {
 
 }
 
-export default ContoService;
+export default QuizzService;
